refactor: project structure refactoring
