+++ pmm-app/package.json
@@ -32,7 +32,6 @@
     "babel-jest": "^24.9.0",
     "copy-webpack-plugin": "^4.5.1",
     "css-loader": "^3.2.0",
-    "enzyme": "^3.10.0",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "jest": "^24.9.0",
     "jest-cli": "^24.9.0",
