+++ pmm-app/src/pmm-add-instance-app-panel/DiscoveryPanel/CredentialsForm.tsx
@@ -1,8 +1,8 @@
 import { Form as FormFinal } from 'react-final-form';
 import React, { ReactElement } from 'react';
 import { useForm } from 'react-final-form-hooks';
-import { InputField } from '../../react-plugins-deps/components/FieldsComponents/Input/Input';
-import { PasswordField } from '../../react-plugins-deps/components/FieldsComponents/Password/Password';
+import { InputField } from '../../react-plugins-deps/components/FormComponents/Input/Input';
+import { PasswordField } from '../../react-plugins-deps/components/FormComponents/Password/Password';
 
 const CredentialsForm = props => {
   return (
