+++ pmm-app/src/pmm-inventory/Inventory.service.ts
