+++ pmm-app/src/pmm-settings/Parts/Diagnostics/Diagnostics.test.tsx
