+++ pmm-app/src/pmm-settings/Parts/Settings/Settings.test.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import ShallowRenderer from 'react-test-renderer/shallow';
 // import renderer from 'react-test-renderer';
 const renderer = new ShallowRenderer();
-jest.mock('../../react-plugins-deps/components/helpers/notification-manager', () => () => ({
+jest.mock('../../../react-plugins-deps/components/helpers/notification-manager', () => () => ({
   showErrorNotification: () => {},
 }));
 // TODO: add correct shallow renderer, we need to have something to render to 2nd level
