+++ pmm-app/src/pmm-settings/Parts/UploadSSH/UploadSSH.test.tsx
@@ -1,23 +1,12 @@
 // Just a stub test
 import React from 'react';
 import renderer from 'react-test-renderer';
-import UploadSSHKey from './UploadSSHKey';
-jest.mock('../../react-plugins-deps/components/helpers/notification-manager', () => () => ({
-  strings: {
-    polish: {
-      agree: 'tak',
-      disagree: 'nie',
-    },
-    malaysian: {
-      agree: 'ya',
-      disagree: 'tidak',
-    },
-  },
-}));
+import UploadSSH from './UploadSSH';
+jest.mock('../../../react-plugins-deps/components/helpers/notification-manager', () => () => ({}));
 describe('Settings Part test', () => {
   it('Upload SSH key renders correct without props', () => {
     const component = renderer.create(
-      <UploadSSHKey
+      <UploadSSH
         settings={{
           ssh_key: 'test_ssh_key',
         }}
