+++ pmm-app/src/react-plugins-deps/components/FormComponents/Checkbox/Checkbox.scss
