+++ pmm-app/src/react-plugins-deps/components/FormComponents/CollapsePanel/CollapsePanel.test.tsx
