+++ pmm-app/src/react-plugins-deps/components/FormComponents/SelectTags/SelectTags.test.tsx
