+++ pmm-app/src/react-plugins-deps/components/FormComponents/Slider/Slider.scss
