+++ pmm-app/src/react-plugins-deps/components/FormComponents/Toggle/Toggle.scss
